requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blinker==1.6.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
decorator==4.4.2
filelock==3.12.0
flatbuffers==23.1.21
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==0.4.1
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.57.1
grpcio==1.42.0
grpcio-status==1.41.1
h5py==3.7.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipython-genutils==0.2.0
jax==0.4.8
jedi==0.18.0
Jinja2==3.1.2
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.2
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
parso==0.8.1
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyu2f==0.1.5
pywin32==300
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.20.0
setuptools==66.0.0
six==1.15.0
sympy==1.11.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.84
urllib3==1.26.15
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.11.2
yarl==1.7.2
zipp==3.15.0
